correcciones

+++ look.domain/entities/proyecto/TarifarioConvenio.cs
@@ -10,7 +10,7 @@ public class TarifarioConvenio
     /// id de Entidad Perfil 
     /// </summary>
     public int? TcPerfilAsignado { get; set; }
-    public int? TcTarifa { get; set; }
+    public double? TcTarifa { get; set; }
     public int TcMoneda { get; set; }
     public int? TcBase { get; set; }
     public int? TcStatus { get; set; }
